Run singularity with subprocess and catch output
Use subprocess instead of execlp, since the latter seemed to cause
problems for failure detection in cluster_utils.
Use this to also pipe the output to files in working_dir.  This way, we
also get the output when running locally (cluster_utils does not store
it in this case), which makes debugging much easier.