hysr_start_robots: Monitor subprocesses
Keep track of the PIDs of the started subprocesses and monitor them.  If
termination is requested via Ctrl+C or any of the subprocesses
terminates on their own, all other processes are killed before existing.

This change makes it easier to detect failures in an automated setup.
It is still possible to run things in the background if needed by using

    hysr_start_robots &