Workaround for cluster_utils 2.4 thinking all runs fail
`cluster.read_params_from_cmdline` not only reads the parameters but
also registers the job at the cluster_utils server as well as
registering a function that is called when the script exists (to tell
the server that the job finished).

Since it is called twice in our setup (once in run_singularity.py and
once in run_learning.py), this resulted in double-reporting which is not
expected by the server and thus resulted in all jobs being considered as
failed.

By removing server info from the arguments, this step is skipped in
run_singularity.py, so communication with the server only happens in
run_learning.py.  While this seems to work, it can only be considered a
rather dirty workaround and should be fixed properly on the long term
(though this will likely require some change in cluster_utils itself).