hp_optim: Run learning multiple times in each job
To reduce the noise in the estimated reward for a given configuration,
run the learning multiple times in each job and use mean reward as
metric for the optimisation.

This is implemented using `cluster.exit_for_resume()`, i.e. for each
iteration of learning the cluster job is restarted.  This way the
duration of the single cluster jobs does not increase and it also makes
it easier to handle the restarts in case of failure more properly.
The "restarts" log file is extended for this to not only include the
number of failed runs but also the rewards of the finished runs and a
"finished runs" counter.

Further changes:

- Use ".json" for restarts log. Since the config file is explicitly
  passed to the hysr scripts now, there is no reason to not use ".json"
  here.
- Make the max. number of retries in case of failures configurable.