hp_optim: Fail if unknown hyperparameter is specified
When creating the config files for a run, check for each hyperparameter
update, if the corresponding parameter actually exists in the config
template.  Raise an error if not.

Previously a typo in the parameter name let to the value silently being
added to the config file with the wrong name, leaving the actual
parameter at its default value.