Add option to split training into multiple cluster jobs
When the "training_iterations" parameter is set to something > 1, the
training is, once finished, marked for continuation.  This means a new
cluster job will be started which loads the model of the previous run
and continues training from there.
This can help saving cluster money if the training takes very long.

For this to work more cleanly, create a "run data" directory for each
job (based on `run_id`).  All run-specific files (config files, training
logs, final models) are stored there.

Also rename "runs" to "trainings" in RestartInfo, to be more clear.

Save JSON files with indentation to make them more human-readable.